+++ lib/src/const.dart
@@ -1,0 +1,55 @@
+part of katana;
+
+/// Class that defines a constant value.
+class Const {
+  /// ID.
+  static const String id = "id";
+
+  /// Universal ID;
+  static const String uid = "uid";
+
+  /// Name.
+  static const String name = "name";
+
+  /// Text.
+  static const String text = "text";
+
+  /// User.
+  static const String user = "user";
+
+  /// Time.
+  static const String time = "time";
+
+  /// Type.
+  static const String type = "type";
+
+  /// Medias.
+  static const String media = "media";
+  
+  /// Category.
+  static const String category = "category";
+  
+  /// Tag.
+  static const String tag = "tag";
+
+  /// Created time.
+  static const String createdTime = "createdTime";
+
+  /// Modified time.
+  static const String modifiedTime = "modifiedTime";
+
+  /// Selection.
+  static const String selection = "selection";
+
+  /// Required.
+  static const String required = "required";
+
+  /// Optional.
+  static const String optional = "optional";
+  
+  /// Question.
+  static const String question = "question";
+  
+  /// Answer.
+  static const String answer = "answer";
+}